README.md, test.c, test.py: reduce and document requirements
The C23 attribute syntax (e.g. [[noreturn]]) is very welcome, but
__attribute__ is likely to be more widely supported and we don't need
any other newer-than-C99 features. So just stick to C99 plus
__attribute__.

(#ifdef-guarding __attribute__ would be possible, but compiling picolog
requires a compiler that understands the system C headers on a *nix
system; I'm not sure that any such compiler exists that doesn't
understand __attribute__. All of gcc/clang/tcc do.)

We don't use anything that's not in Python-3.10, so just require that
(I'm not going to test against unsupported versions).